refactor(omp): rename OMQ to OMP (Oh My Paw) to avoid name conflict
