+++ plugins/bundle/omp_workflows/plugin.json
@@ -1,9 +1,9 @@
 {
-  "id": "omq-workflows",
-  "name": "OMQ Workflows",
+  "id": "omp-workflows",
+  "name": "OMP Workflows",
   "version": "0.1.0",
   "type": "bundle",
-  "description": "Oh My QwenPaw workflow modes: Autopilot, Ralph, UltraQA, Ultrawork, Team",
+  "description": "Oh My Paw workflow modes: Autopilot, Ralph, UltraQA, Ultrawork, Team",
   "author": "QwenPaw Team",
   "entry": {
     "backend": "plugin.py"
