+++ plugins/bundle/omp_workflows/plugin.py
@@ -1,16 +1,16 @@
 # -*- coding: utf-8 -*-
-"""OMQ Workflows plugin — registers five workflow AgentModes."""
+"""OMP Workflows plugin — registers five workflow AgentModes."""
 
 import logging
 from pathlib import Path
 
-logger = logging.getLogger("qwenpaw").getChild("plugin.omq_workflows")
+logger = logging.getLogger("qwenpaw").getChild("plugin.omp_workflows")
 
 _PLUGIN_DIR = Path(__file__).parent
 
 
-class OMQWorkflowsPlugin:
-    """Plugin entry point for Oh My QwenPaw workflow modes."""
+class OMPWorkflowsPlugin:
+    """Plugin entry point for Oh My Paw workflow modes."""
 
     def register(self, api) -> None:
         from .autopilot.mode import AutopilotMode
@@ -27,7 +27,7 @@ class OMQWorkflowsPlugin:
             TeamMode,
         ):
             api.register_mode(mode_cls)
-            logger.info("Registered OMQ mode: %s", mode_cls.name)
+            logger.info("Registered OMP mode: %s", mode_cls.name)
 
         api.register_skill_provider(
             skills_dir=_PLUGIN_DIR / "skills",
@@ -35,4 +35,4 @@ class OMQWorkflowsPlugin:
         )
 
 
-plugin = OMQWorkflowsPlugin()
+plugin = OMPWorkflowsPlugin()
