+++ plugins/bundle/omp_workflows/shared/role_prompts.py
