+++ plugins/bundle/omp_workflows/team/mode.py
@@ -1,9 +1,10 @@
 # -*- coding: utf-8 -*-
-"""UltraQAMode — QA cycle engine with 3-agent collaboration."""
+"""TeamMode — multi-agent collaboration pipeline."""
 
 from __future__ import annotations
 
 import logging
+import re
 import shlex
 from typing import TYPE_CHECKING, Optional
 
@@ -15,32 +16,32 @@ from qwenpaw.runtime.slash_command_registry import CommandSpec
 if TYPE_CHECKING:
     from typing import Any
 
-    from .gate import UltraQAGate
+    from .gate import TeamPipelineGate
 
 logger = logging.getLogger(__name__)
 
 _HELP = (
-    "**UltraQA** — automated QA cycle engine\n\n"
-    "Usage:\n"
-    "  `/ultraqa [--tests|--build|--lint|--typecheck|--custom \"cmd\"]"
-    " [--interactive] [target]`\n\n"
-    "Runs repeated QA cycles: check → diagnose → fix → re-check.\n"
-    "Stops when all checks pass or max cycles reached."
+    "**Team** — multi-agent collaboration pipeline\n\n"
+    "Usage: `/team [N:role] <task>`\n\n"
+    "Examples:\n"
+    "  `/team 3:executor Implement authentication`\n"
+    "  `/team ralph Build the REST API`\n\n"
+    "Phases: plan -> prd -> exec -> verify -> fix (retry)"
 )
 
 
-class UltraQAMode(AgentMode):
-    """AgentMode for the UltraQA workflow."""
+class TeamMode(AgentMode):
+    """AgentMode for the Team pipeline."""
 
-    name = "ultraqa"
+    name = "team"
 
     def __init__(self) -> None:
-        self._gate: UltraQAGate | None = None
+        self._gate: TeamPipelineGate | None = None
 
     def commands(self) -> list[CommandSpec]:
         return [
             CommandSpec(
-                name="ultraqa",
+                name="team",
                 handler=self._handler,
                 category="builtin",
                 help_text=_HELP,
@@ -52,7 +53,7 @@ class UltraQAMode(AgentMode):
         super().setup(workspace)
         from qwenpaw.loop.gates import StopHandler, StopHandlerRegistration
 
-        from .gate import UltraQAGate as _G
+        from .gate import TeamPipelineGate as _G
 
         handler = StopHandler()
         gate = _G()
@@ -65,11 +66,11 @@ class UltraQAMode(AgentMode):
                 plugins.stop_handlers = []
             plugins.stop_handlers.append(
                 StopHandlerRegistration(
-                    plugin_id="__omq_ultraqa__",
+                    plugin_id="__omp_team__",
                     handler=handler,
                     priority=0,
-                    name="ultraqa-stop-handler",
-                    scope="omq-ultraqa",
+                    name="team-stop-handler",
+                    scope="omp-team",
                 ),
             )
 
@@ -81,8 +82,9 @@ class UltraQAMode(AgentMode):
             return _info(_HELP)
 
         parsed = _parse_args(args)
-        if parsed is None:
-            return _info("Invalid arguments. " + _HELP)
+        task = parsed["task"]
+        if len(task) < 5:
+            return _info("Please provide a task description.\n\n" + _HELP)
 
         workspace_dir = getattr(ctx, "workspace_dir", None)
         if not workspace_dir:
@@ -90,57 +92,51 @@ class UltraQAMode(AgentMode):
 
         from pathlib import Path
 
-        loop_dir = self._gate.activate_for_qa(
+        loop_dir = self._gate.activate_for_team(
             workspace_dir=Path(workspace_dir),
-            goal_type=parsed["goal_type"],
-            custom_cmd=parsed["custom_cmd"],
-            interactive=parsed["interactive"],
+            agent_count=parsed["agent_count"],
+            agent_role=parsed["agent_role"],
         )
 
         prompt = (
-            f"UltraQA activated. Goal: {parsed['goal_type']}.\n"
+            f"Team pipeline activated.\n"
+            f"Task: {task}\n"
+            f"Workers: {parsed['agent_count']}, Role: {parsed['agent_role']}\n"
             f"State directory: {loop_dir}\n"
-            f"Read {loop_dir}/state.json and begin the QA cycle."
+            f"Phase: plan — explore the codebase and create a task breakdown."
         )
         _rewrite(ctx, prompt)
-        logger.info("UltraQA started: %s", loop_dir)
+        logger.info("Team started: %s", loop_dir)
         return None
 
 
-def _parse_args(raw: str) -> dict | None:
-    """Parse /ultraqa arguments."""
+_TEAM_SPEC_RE = re.compile(r"^(\d+):(\w[\w-]*)$")
+
+
+def _parse_args(raw: str) -> dict:
     try:
         tokens = shlex.split(raw)
     except ValueError:
-        return None
+        return {"task": raw, "agent_count": 3, "agent_role": "executor"}
 
-    goal_type = "tests"
-    custom_cmd = ""
-    interactive = False
+    agent_count = 3
+    agent_role = "executor"
+    task_parts: list[str] = []
 
-    i = 0
-    while i < len(tokens):
-        t = tokens[i]
-        if t == "--tests":
-            goal_type = "tests"
-        elif t == "--build":
-            goal_type = "build"
-        elif t == "--lint":
-            goal_type = "lint"
-        elif t == "--typecheck":
-            goal_type = "typecheck"
-        elif t == "--interactive":
-            interactive = True
-        elif t == "--custom" and i + 1 < len(tokens):
-            goal_type = "custom"
-            i += 1
-            custom_cmd = tokens[i]
-        i += 1
+    for i, t in enumerate(tokens):
+        m = _TEAM_SPEC_RE.match(t)
+        if m and i == 0:
+            agent_count = int(m.group(1))
+            agent_role = m.group(2)
+        elif t in ("executor", "ralph") and i == 0:
+            agent_role = t
+        else:
+            task_parts.append(t)
 
     return {
-        "goal_type": goal_type,
-        "custom_cmd": custom_cmd,
-        "interactive": interactive,
+        "task": " ".join(task_parts),
+        "agent_count": max(1, min(agent_count, 10)),
+        "agent_role": agent_role,
     }
